ARCC_Dev-main/ARCC_Consultant/requirements.txt
djangorestframework 
django
openai
tiktoken
python-dotenv
markdown
pandas
djangorestframework-simplejwt